+++ src/main/java/br/com/david/orcamento/service/ProgramaService.java
@@ -1,6 +1,8 @@
 package br.com.david.orcamento.service;
 
+import br.com.david.orcamento.model.LancamentoModel;
 import br.com.david.orcamento.model.ProgramaModel;
+import br.com.david.orcamento.repository.LancamentoRepository;
 import br.com.david.orcamento.repository.ProgramaRepository;
 import br.com.david.orcamento.rest.components.DataFormato;
 import br.com.david.orcamento.rest.dto.ProgramaDTo;
@@ -24,7 +26,8 @@ public class ProgramaService {
 
     @Autowired
     ProgramaRepository programaRepository;
-
+    @Autowired
+    LancamentoRepository lancamentoRepository;
     @Autowired
     ModelMapper modelMapper;
 
@@ -97,7 +100,16 @@ public class ProgramaService {
         {
             if (programaRepository.existsById(id))
             {
-                programaRepository.deleteById(id);
+                List<LancamentoModel> lancamentos = lancamentoRepository.findAll();
+
+                for (LancamentoModel lancamento : lancamentos ) {
+                    if (!(lancamento.getId_programa().equals(id))){
+                        programaRepository.deleteById(id);
+                    } else {
+                        throw new DataIntegrityException("Este programa esta contido em um lançamento!");
+                    }
+                }
+
             } else
             {
                 throw new ObjectNotFoundException("Códido de ID: " + id + " não encontrado!");
